requirements.txt
boto3
requests
Pillow
tqdm
tensorflow
opencv-python-headless
matplotlib
numpy
